requirements.txt
Django
Pillow
djangorestframework
djangorestframework-simplejwt
django-cors-headers
drf-yasg